Modifico repositorios

+++ app/src/main/java/com/example/appdelclima/repository/modelos/Ciudad.kt
@@ -3,8 +3,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Ciudad(
     val name: String,
-    val lat: Double,
-    val lon: Double,
+    val lat: Float,
+    val lon: Float,
     val country: String,
     val state: String? = null
 )
